+++ packages/core/cli/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/cli@0.3.12...@kong-ui-public/cli@0.4.0) (2023-02-15)
+
+
+### ✨ Features
+
+* **misc-widgets:** create new package ([#134](https://github.com/Kong/public-ui-components/issues/134)) ([0cf32a1](https://github.com/Kong/public-ui-components/commit/0cf32a16f2c733947513754de9482e41a1d0afbf))
+
+
+
+
+
+## [0.3.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/cli@0.3.11...@kong-ui-public/cli@0.3.12) (2023-02-15)
+
+
+### ♻️ Chores
+
+* **cli:** inject current year ([6804d0a](https://github.com/Kong/public-ui-components/commit/6804d0ad333ca9a2041392eb9026794464a112a9))
+
+
+
+
+
+## [0.3.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/cli@0.3.10...@kong-ui-public/cli@0.3.11) (2023-02-11)
+
+
+### 📚 Documentation
+
+* **cli:** add documentation ([#137](https://github.com/Kong/public-ui-components/issues/137)) ([0e41527](https://github.com/Kong/public-ui-components/commit/0e415275c6fc1d5f3d2444dd74fc037fb747c7c0))
+
+
+
+
+
 ## [0.3.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/cli@0.3.9...@kong-ui-public/cli@0.3.10) (2023-02-07)